Localhost simulation 45 (#46)
* don't build binary for localhost
