+++ simul/simul.go
@@ -16,7 +16,26 @@ import (
 	"os"
 
 	"github.com/dedis/onet/log"
+	"github.com/dedis/onet/simul/platform"
 )
+
+// The address of this conode - if there is only one conode in the config
+// file, it will be derived from it automatically
+var conodeAddress string
+
+// ip addr of the logger to connect to
+var monitorAddress string
+
+// Simul is != "" if this node needs to start a simulation of that protocol
+var simul string
+
+// Initialize before 'init' so we can directly use the fields as parameters
+// to 'Flag'
+func init() {
+	flag.StringVar(&conodeAddress, "address", "", "our address to use")
+	flag.StringVar(&simul, "simul", "", "start simulating that protocol")
+	flag.StringVar(&monitorAddress, "monitor", "", "remote monitor")
+}
 
 // Start has to be called by the main-file that imports the protocol and/or the
 // service. If a user calls the simulation-file, `simul` is empty, and the
@@ -40,6 +59,7 @@ func Start(rcs ...string) {
 	if simul == "" {
 		startBuild()
 	} else {
-		simulate()
+		err := platform.Simulate(conodeAddress, simul, monitorAddress)
+		log.ErrFatal(err)
 	}
 }
